refactor: 파일 구조 변경 및 공백 수정
