+++ Diary/View/DiaryTableViewCell.swift
@@ -10,7 +10,6 @@ import UIKit
 final class DiaryTableViewCell: UITableViewCell {
     private let titleLabel: UILabel = {
         let label = UILabel()
-        
         label.font = .preferredFont(forTextStyle: .title2)
         label.adjustsFontForContentSizeCategory = true
         label.translatesAutoresizingMaskIntoConstraints = false
@@ -20,7 +19,6 @@ final class DiaryTableViewCell: UITableViewCell {
     
     private let infoLabel: UILabel = {
         let label = UILabel()
-        
         label.font = .preferredFont(forTextStyle: .footnote)
         label.adjustsFontForContentSizeCategory = true
         label.translatesAutoresizingMaskIntoConstraints = false
@@ -30,7 +28,6 @@ final class DiaryTableViewCell: UITableViewCell {
     
     private let diaryInfoStackView: UIStackView = {
         let stackView = UIStackView()
-        
         stackView.axis = .vertical
         stackView.spacing = 8
         stackView.translatesAutoresizingMaskIntoConstraints = false
@@ -45,7 +42,6 @@ final class DiaryTableViewCell: UITableViewCell {
     
     override func prepareForReuse() {
         super.prepareForReuse()
-        
         titleLabel.text = nil
         infoLabel.text = nil
     }
@@ -57,9 +53,7 @@ final class DiaryTableViewCell: UITableViewCell {
     private func configureCell() {
         diaryInfoStackView.addArrangedSubview(titleLabel)
         diaryInfoStackView.addArrangedSubview(infoLabel)
-        
-        self.contentView.addSubview(diaryInfoStackView)
-        
+        contentView.addSubview(diaryInfoStackView)
         NSLayoutConstraint.activate([
             diaryInfoStackView.topAnchor.constraint(equalTo: self.contentView.topAnchor,
                                                     constant: 8),
@@ -75,8 +69,8 @@ final class DiaryTableViewCell: UITableViewCell {
     func configureLabel(diaryData: DiaryData) {
         guard let title = diaryData.title else { return }
         let dateText = DateManger.shared.convertToDate(fromDouble: diaryData.createDate)
-        
         titleLabel.text = title
+        
         if let body = diaryData.body {
             infoLabel.text = dateText + "  " + body
         } else {
@@ -91,9 +85,7 @@ extension UILabel {
         let fullText = self.text ?? ""
         let range = (fullText as NSString).range(of: targetString)
         let attributedString = NSMutableAttributedString(string: fullText)
-        
         attributedString.addAttribute(.font, value: UIFont.preferredFont(forTextStyle: .callout), range: range)
-        
         self.attributedText = attributedString
     }
 }
