+++ DiaryData+CoreDataProperties.swift
@@ -9,7 +9,6 @@ import Foundation
 import CoreData
 
 extension DiaryData {
-
     @nonobjc public class func fetchRequest() -> NSFetchRequest<DiaryData> {
         return NSFetchRequest<DiaryData>(entityName: "DiaryData")
     }
@@ -18,7 +17,6 @@ extension DiaryData {
     @NSManaged public var title: String?
     @NSManaged public var body: String?
     @NSManaged public var createDate: Double
-
 }
 
 extension DiaryData: Identifiable { }
